Cars/core/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-admin-list-filter-dropdown==1.0.3
django-ckeditor==6.5.1
django-extensions==3.2.1
django-import-export==3.1.0
django-js-asset==2.0.0
djangorestframework==3.12.1
et-xmlfile==1.1.0
executing==1.2.0
Faker==17.6.0
faker-vehicle==0.2.0
ipython==8.11.0
jedi==0.18.2
MarkupPy==1.14
matplotlib-inline==0.1.6
odfpy==1.4.1
openpyxl==3.1.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tablib==3.3.0
traitlets==5.9.0
tzdata==2022.7
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
